[model] fix: change SkinVertex::bones to glm::ivec4..

+++ lib/nw/model/Mdl.hpp
@@ -420,7 +420,7 @@ struct SkinVertex {
     glm::vec2 tex_coords{};
     glm::vec3 normal{};
     glm::vec4 tangent{};
-    glm::i16vec4 bones{};
+    glm::ivec4 bones{};
     glm::vec4 weights{};
 };
